+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Make sure that `get_items` is backwards compatible ([#1139](https://github.com/stac-utils/pystac/pull/1139))
 - Make `_repr_html_` look like `_repr_json_` output ([#1142](https://github.com/stac-utils/pystac/pull/1142))
 - Improved error message when `.ext` is called on a Collection ([#1157](https://github.com/stac-utils/pystac/pull/1157))
+- `add_child` and `add_item` return a Link object instead of None [#1159](https://github.com/stac-utils/pystac/issues/1159)
 
 ### Deprecated